Fix bugs in pkg/fileutil

+++ localpaas_app/pkg/fileutil/cert_file.go
@@ -23,29 +23,36 @@ func WriteCerts(cert, privateKey []byte, saveDir, certFile, keyFile string, over
 	certPath := filepath.Join(saveDir, certFile)
 	keyPath := filepath.Join(saveDir, keyFile)
 
+	var writtenCert, writtenKey bool
 	defer func() {
 		if err == nil {
 			return
 		}
-		// Try to remove all created files
-		_ = os.Remove(keyPath)
-		_ = os.Remove(certPath)
+		// Try to remove ONLY the files we just created during this run
+		if writtenKey {
+			_ = os.Remove(keyPath)
+		}
+		if writtenCert {
+			_ = os.Remove(certPath)
+		}
 	}()
 
-	if certPath != "" {
+	if certFile != "" {
 		if overwrite || !gofn.Head(FileExists(certPath, true)) {
 			err = os.WriteFile(certPath, cert, certFileMode)
 			if err != nil {
 				return apperrors.New(err).WithMsgLog("failed to write cert file to %s", certPath)
 			}
+			writtenCert = true
 		}
 	}
-	if keyPath != "" {
+	if keyFile != "" {
 		if overwrite || !gofn.Head(FileExists(keyPath, true)) {
 			err = os.WriteFile(keyPath, privateKey, certFileMode)
 			if err != nil {
 				return apperrors.New(err).WithMsgLog("failed to write key file to %s", keyPath)
 			}
+			writtenKey = true
 		}
 	}
